refactor: reorganize rich editor plugin imports and implement link functionality
