--- src/components/rich-editor/plugins/README.md
@@ -1,1 +1,0 @@
-
