+++ src/components/rich-editor/plugins/mentions/components/mention-picker.tsx
