+++ src/components/rich-editor/plugins/mentions/mention-node.tsx
